rfct: Move multiple source files to subpackages
